// types/domain.ts
export type Artisan = {
  id: string;
  name: string;
  specialty: string;
  neighborhood: string;
  rating: number;
  distance: string;
  categories: string[];
};

export type Garment = {
  id: string;
  name: string;
  brand: string;
  size: string;
  price: number;
  condition: string;
  category: string;
  emoji: string;
  color: string;
  seller: string;
  neighborhood: string;
  rating: number;
  description: string;
};